jwt summd up with role

+++ src/main/java/com/springboot/microservices/user/dao/UserDao.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.springboot.microservices.user.model.User;
 
 @Repository
-public interface UserDao extends JpaRepository<User, Integer>, CrudRepository<User, Integer>{
+public interface UserDao extends JpaRepository<User, Long>, CrudRepository<User, Long>{
 
 	Optional<User> findByUserName(String username);
